+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.0] 2020-05-09
+### Added
+- User profile picture to comment
+- User username to comment
+- Timestamp to comment
+- All of the dating functionality
+
 ## [0.3.1] 2020-05-09
 ### Added
 - Tests for all Routers